#2 code cleanups - renaming
