+++ src/app/components/JiraItemDetails.tsx
